--HAS A SEGFAULT--
Added structures which will be our system to build game objects out of voxels (Such as a player)

Added mesh merging and implementation of monotone meshing to work with the structure system so it doesn't reduce performance

Added GeometryManager so the game engine (Structures in this case) can update things such as the color and the vertices for the user and thus we reduce the amount of boilerplate code.

Added string based methods for some of our get functions.

Minor fixes.